23002170110139_SAJJAD/projeXt/newPythonOllama/ollama-web-ui/ollama_web_ui/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
requests==2.31.0
Pillow==10.1.0